Rename to client_frontend
